chore: rename io.rpg.torefract -> io.rpg.torefact
